+++ R/plot_egg_slopes.R
@@ -8,6 +8,7 @@
 #'
 #' @examples
 #' library(ggplot2)
+#' library(eggla)
 #' data("bmigrowth")
 #' res <- egg_model(
 #'   formula = log(bmi) ~ age,
@@ -18,15 +19,7 @@
 #' plot_egg_slopes(
 #'   fit = res,
 #'   period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17)
-#' ) +
-#'   theme_minimal() +
-#'   theme(
-#'     axis.line = element_line(colour = "grey20", size = rel(1)),
-#'     axis.ticks = element_line(colour = "grey20"),
-#'     panel.grid.minor.x = element_blank(),
-#'     panel.grid.minor.y = element_blank(),
-#'     panel.grid.major.x = element_blank()
-#'   )
+#' )
 plot_egg_slopes <- function(fit, period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17), knots = c(1, 8, 12)) {
   params <- variable <- yend <- pred_period <- end <- NULL # no visible binding for global variable from data.table
   patterns <- slope <- .data <- start <- NULL # no visible binding for global variable from data.table